+++ R/combineFeatures.R
@@ -5,9 +5,10 @@ combineFeatures <- function(object,
                                 "weighted.mean",
                                 "sum",
                                 "medpolish",
+                                "robust",
                                 "iPQF",
                                 "NTR"),
-                            fcol,                            
+                            fcol,
                             redundancy.handler = c("unique", "multiple"),
                             cv = TRUE,
                             cv.norm = "sum",
@@ -22,6 +23,11 @@ combineFeatures <- function(object,
         stopifnot(fcol %in% fvarLabels(object))
         groupBy <- fData(object)[, fcol]
     }
+    if (anyNA(object)) {
+        msg <- "Your data contains missing values. Please read the relevant section in the combineFeatures manual page for details the effects of missing values on data aggregation."
+        message(paste(strwrap(msg), collapse = "\n"))
+    }
+
     if (is.list(groupBy)) {
         if (length(groupBy) != nrow(object))
             stop("'length(groupBy)' must be equal to 'nrow(object)': ",
@@ -88,8 +94,22 @@ combineFeaturesV <- function(object,   ## MSnSet
                              ) {
     groupBy <- as.character(groupBy)
     if (cv) {
+        ## New cv feature variable names
+        cvfvars <- paste("CV", sampleNames(object), sep = ".")
+        ## If not already present, use as is (i.e no suffix needed)
+        if (!any(cvfvars %in% fvarLabels(object))) {
+            .suffix <- NULL
+        } else {
+            ## Add a numeric suffix that isn't already in use
+            .suffix <- 0
+            while (any(cvfvars %in% fvarLabels(object))) {
+                .suffix <- .suffix + 1
+                cvfvars <- paste(cvfvars, .suffix, sep = ".")
+            }
+        }
         cv.mat <- featureCV(object, groupBy = groupBy,
-                            norm = cv.norm)
+                            norm = cv.norm,
+                            suffix = .suffix)
     }
     n1 <- nrow(object)
     ## !! order of features in matRes is defined by the groupBy factor !!
@@ -156,6 +176,8 @@ combineMatrixFeatures <- function(matr,    ## matrix
                                          medpol <- medpolish(x, trace.iter = verbose, ...)
                                          return(medpol$overall + medpol$col)
                                      })
+        } else if (fun == "robust") {
+            summarisedFeatures <- by(matr, groupBy, robustSummary, ...)
         } else if (fun == "weighted.mean") {
             ## Expecting 'w' argument
             args <- list(...)
@@ -193,7 +215,6 @@ combineMatrixFeatures <- function(matr,    ## matrix
     }
     return(do.call(rbind, as.list(summarisedFeatures)))
 }
-
 
 ##' This function evaluates the variability within all protein group
 ##' of an \code{MSnSet}. If a protein group is composed only of a
@@ -246,4 +267,64 @@ aggvar <- function(object, groupBy, fun) {
     ans <- cbind(agg_dist = d, nb_feats = nr)
     attr(ans, "agg_dist") <- fun
     ans
+}
+
+##' This function calculates the robust summarisation for each feature
+##' (protein). Note that the function assumes that the intensities in
+##' input `e` are already log-transformed.
+##'
+##' @title Calculate robust expression summary
+##' @param e A feature (peptide or spectra) by sample `matrix`
+##'     containing the expression data.
+##' @param ... Additional parameters passed to `MASS::rlm`
+##' @return `numeric()` vector of length `length(expression)` with
+##'     robust summarised values.
+##' @author Adriaan Sticker, Sebastian Gibb and Laurent Gatto
+##' @md
+##' @noRd
+robustSummary <- function(e, residuals = FALSE, ...) {
+    ## If there is only one 1 peptide for all samples return
+    ## expression of that peptide
+    if (nrow(e) == 1L) return(e)
+
+    ## remove missing values
+    p <- !is.na(e)
+    expression <- e[p] ## expression becomes a vector
+    sample <- rep(colnames(e), each = nrow(e))[p]
+    feature <- rep(rownames(e), times = ncol(e))[p]
+
+    ## model.matrix breaks on factors with 1 level so make vector of
+    ## ones (intercept).
+    if (length(unique(sample)) == 1L) sample <- rep(1, length(sample))
+
+    ## Sum contrast on peptide level so sample effect will be mean
+    ## over all peptides instead of reference level.
+    X <- stats::model.matrix(~ -1 + sample + feature,
+                             contrasts.arg = list(feature = 'contr.sum'))
+    ## MASS::rlm breaks on singulare values.
+    ## - Check with base lm if singular values are present.
+    ## - If so, these coefficients will be zero, remove this collumn
+    ##   from model matrix
+    ## - Rinse and repeat on reduced modelmatrx till no singular
+    ##   values are present
+    repeat {
+        fit <- stats::.lm.fit(X, expression)
+        id <- fit$coefficients != 0
+        X <- X[ , id, drop = FALSE]
+        if (!any(!id)) break
+    }
+    ## Last step is always rlm: calculate estimated effects effects as
+    ## summarised values
+    fit <- MASS::rlm(X, expression, ...)
+
+    sampleid <- seq_along(unique(sample))
+    ## This will be needed for NUSE-type of quality control, but will
+    ## need to check for missing data as below.
+    ## se <- unique(summary(fit)$coefficients[sampleid, 'Std. Error'])
+
+    ## Put NA for the samples without any expression value
+    present <- as.logical(colSums(p))
+    res <- rep(NA, length(present))
+    res[present] <- fit$coefficients[sampleid]
+    res
 }