+++ R/readMSData.R
@@ -14,7 +14,7 @@
 ##' files. This can be achieved by setting `mode = "onDisk"`. The
 ##' function returns then an [OnDiskMSnExp-class] object instead of a
 ##' [MSnExp-class] object.
-##' 
+##'
 ##' @title Imports mass-spectrometry raw data files as 'MSnExp'
 ##'     instances.
 ##' @note `readMSData` uses `normalizePath` to replace relative with
@@ -32,10 +32,12 @@
 ##' @param verbose Verbosity flag. Default is to use
 ##'     [isMSnbaseVerbose()].
 ##' @param centroided. A `logical`, indicating whether spectra are
-##'     centroided or not. Default is `NA`. In `onDisk`, it can also
-##'     be set for different MS levels by a vector of logicals, where
-##'     the first element is for MS1, the second element is for MS2,
-##'     ... See [OnDiskMSnExp-class] for an example.
+##'     centroided or not. Default is `NA` in which case the information
+##'     is extracted from the raw file (for mzML or mzXML files). In
+##'     `onDisk`, it can also be set for different MS levels by a
+##'     vector of logicals, where the first element is for MS1, the
+##'     second element is for MS2, ... See [OnDiskMSnExp-class] for
+##'     an example.
 ##' @param smoothed. A `logical` indicating whether spectra already
 ##'     smoothed or not. Default is `NA`.
 ##' @param cache. Numeric indicating caching level. Default is 0 for
@@ -66,18 +68,45 @@ readMSData <- function(files, pdata = NULL, msLevel. = NULL,
     ##   path. That fixes possible problems on Windows with SNOW parallel
     ##   processing and also proteowizard problems on unis system with ~ paths.
     files <- normalizePath(files)
-    if (mode == "inMemory") {
-        if (is.null(msLevel.)) msLevel. <- 2L
-        readInMemMSData(files, pdata = pdata, msLevel. = msLevel.,
+    .hasSpecs <- hasSpectra(files)
+    suppressWarnings(.hasChroms <- hasChromatograms(files))
+    if (any(!.hasSpecs)) {
+        msg1 <- paste0("Dropping ", sum(!.hasSpecs),
+                       " file(s) without any spectra: ",
+                       paste(basename(files[!.hasSpecs]),
+                             collapse = ", "), ". ")
+        if (all(.hasChroms[!.hasSpecs]))
+            msg2 <- "They/it contain(s) chromatograms and can be read with `readSRMData()`."
+        else
+            msg2 <- paste0("File(s) ",
+                           paste(basename(files[!.hasSpecs & .hasChroms]),
+                                 collapse = ", "),
+                           "contain(s) chromatograms that can be read with `readSRMData`.")
+        warning(paste0(msg1, msg2))
+    }
+    files <- files[.hasSpecs]
+    if (!length(files)) {
+        process <- new("MSnProcess",
+                       processing = paste("No data loaded:", date()))
+        if (mode == "inMemory")
+            res <- new("MSnExp",
+                       processingData = process)
+        else res <- new("OnDiskMSnExp",
+                        processingData = process)
+    } else {
+        if (mode == "inMemory") {
+            if (is.null(msLevel.)) msLevel. <- 2L
+            res <- readInMemMSData(files, pdata = pdata, msLevel. = msLevel.,
                         verbose = verbose, centroided. = centroided.,
                         smoothed. = smoothed., cache. = cache.)
-    } else { ## onDisk
-        readOnDiskMSData(files = files, pdata = pdata,
-                         msLevel. = msLevel., verbose = verbose,
-                         centroided. = centroided.,
-                         smoothed. = smoothed.)
+        } else { ## onDisk
+            res <- readOnDiskMSData(files = files, pdata = pdata,
+                                    msLevel. = msLevel., verbose = verbose,
+                                    centroided. = centroided.,
+                                    smoothed. = smoothed.)
+        }
     }
-
+    res
 }
 
 
@@ -111,6 +140,10 @@ readInMemMSData <- function(files, pdata, msLevel., verbose,
         msdata <- .openMSfile(f)
         .instrumentInfo <- c(.instrumentInfo, list(instrumentInfo(msdata)))
         fullhd <- mzR::header(msdata)
+        ## Issue #325: get centroided information from file, but overwrite if
+        ## specified with centroided. parameter.
+        if (!is.na(centroided.))
+            fullhd$centroided <- as.logical(centroided.)
         spidx <- which(fullhd$msLevel == msLevel.)
         ## increase vectors as needed
         ioncount <- c(ioncount, numeric(length(spidx)))
@@ -142,7 +175,7 @@ readInMemMSData <- function(files, pdata, msLevel., verbose,
                           mz = .p[, 1],
                           intensity = .p[, 2],
                           fromFile = as.integer(filen),
-                          centroided = as.logical(centroided.),
+                          centroided = as.logical(hd$centroided),
                           smoothed = as.logical(smoothed.),
                           polarity = as.integer(pol))
                 ## peaksCount
@@ -188,7 +221,7 @@ readInMemMSData <- function(files, pdata, msLevel., verbose,
                           mz = .p[, 1],
                           intensity = .p[, 2],
                           fromFile = as.integer(filen),
-                          centroided = as.logical(centroided.),
+                          centroided = as.logical(hd$centroided),
                           smoothed = as.logical(smoothed.),
                           polarity = as.integer(hd$polarity))
                 ## peaksCount