+++ tests/testthat/test_centroided.R
@@ -2,19 +2,15 @@ test_that("centroided accessor with/without na.fail", {
     x <- tmt_erwinia_in_mem_ms2
     x2 <- tmt_erwinia_on_disk
     ## in-mem, single spectrum
-    expect_true(is.na(centroided(x[[1]])))
-    expect_error(centroided(x[[1]], na.fail = TRUE))
+    expect_true(centroided(x[[1]]))
     ## in-mem, experiment
-    expect_true(all(is.na(centroided(x))))
-    expect_error(centroided(x, na.fail = TRUE))
+    expect_true(all(centroided(x)))
     ## on-disk, single spectrum
-    expect_true(is.na(centroided(x2[[1]])))
-    expect_error(centroided(x2[[1]], na.fail = TRUE))
+    expect_true(!centroided(x2[[1]]))
+    expect_true(centroided(x2[[2]]))
     ## on-disk, experiment
-    expect_true(all(is.na(centroided(x2))))
-    expect_error(centroided(x2, na.fail = TRUE))
-    fData(x2)$centroided[1:(length(x2) - 1)] <- TRUE
-    expect_error(centroided(x2, na.fail = TRUE))
+    expect_true(all(!centroided(filterMsLevel(x2, 1))))
+    expect_true(all(centroided(filterMsLevel(x2, 2))))
 })
 
 
@@ -27,14 +23,13 @@ test_that("isCentroidedFromFile", {
     cnt2 <- isCentroidedFromFile(tmt_erwinia_on_disk_ms2)
     expect_identical(names(cnt1), featureNames(tmt_erwinia_on_disk_ms1))
     ##
-    ## multiple files
-    fls <- msdata::proteomics(full.names = TRUE, pattern = "TMT_Erwinia_")
-    x <- readMSData(fls, mode = "onDisk")
-    cnt <- isCentroidedFromFile(x)
-    expect_identical(names(cnt), featureNames(x))
+    ## multiple files; use the microtofq files
+    cnt <- isCentroidedFromFile(microtofq_on_disk)
+    expect_identical(names(cnt), featureNames(microtofq_on_disk))
     ##
-    ## subsetting    
-    k <- sort(sample(length(x), 10))
-    xx <- x[k]
-    expect_identical(isCentroidedFromFile(xx), cnt[k])    
+    ## subsetting
+    set.seed(123) ## see issue #338
+    k <- sort(sample(length(microtofq_on_disk), 10))
+    xx <- microtofq_on_disk[k]
+    expect_identical(isCentroidedFromFile(xx), cnt[k])
 })